update test location and add fix
