+++ packages/model/addon/-private/model.js
@@ -118,8 +118,10 @@ function computeOnce(target, key, desc) {
   @uses DeprecatedEvented
 */
 class Model extends EmberObject {
-  init(...args) {
-    super.init(...args);
+  init(options = {}) {
+    const createProps = options._createProps;
+    delete options._createProps;
+    super.init(options);
 
     if (DEBUG) {
       if (!this._internalModel) {
@@ -132,6 +134,7 @@ class Model extends EmberObject {
     if (CUSTOM_MODEL_CLASS) {
       this.___recordState = new RecordState(this);
     }
+    this.setProperties(createProps);
   }
 
   /**
@@ -2075,6 +2078,7 @@ class Model extends EmberObject {
 // the values initialized during create to `setUnknownProperty`
 Model.prototype._internalModel = null;
 Model.prototype.store = null;
+Model.prototype._createProps = null;
 
 if (HAS_DEBUG_PACKAGE) {
   /**
